# licne-finansije/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Veb aplikacija za pregled licnih finansija",
        "description": "API za upravljanje ličnim finansijama (autentifikacija, transakcije, kategorije, budžeti, finansijski ciljevi, podsetnici, dokumenta i admin statistike)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
...
        "/api/password/forgot": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Slanje linka za resetovanje lozinke",
                "description": "Prima email adresu korisnika, generiše privremeni token za resetovanje i šalje e-mail sa jedinstvenim linkom (npr. preko Mailtrap-a).",
                "operationId": "authSendResetLink",
                "requestBody": {
                    "description": "Email adresa korisnika koji je zaboravio lozinku",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "korisnik@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zahtev je uspešno obrađen i e-mail je poslat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Poslat link za resetovanje lozinke na email adresu ako postoji u sistemu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validacija nije prošla (loš format email adrese)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Neuspesna validacija email adrese"
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email field must be a valid email address."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Korisnik sa tim email-om ne postoji u bazi podataka",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Korisnik sa datom email adresom ne postoji"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/password/reset": {
            "post": {
                "tags": [
                    "Autentifikacija"
                ],
                "summary": "Resetovanje lozinke pomoću tokena",
                "description": "Prima email, token koji je poslat na mejl, kao i novu lozinku (uz potvrdu lozinke). Proverava validnost i istek tokena (60 minuta), pa ažurira lozinku u bazi podataka.",
                "operationId": "authResetPassword",
                "requestBody": {
                    "description": "Podaci za resetovanje lozinke",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "korisnik@example.com"
                                    },
                                    "token": {
                                        "type": "string",
                                        "example": "a1b2c3d4e5f6..."
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NovaLozinka123",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "NovaLozinka123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lozinka je uspešno promenjena",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lozinka uspesno resetovana"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token nije validan ili je istekao",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token za resetovanje lozinke je istekao. Posaljite novi zahtev za resetovanje lozinke."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validacija nije prošla (npr. lozinke se ne poklapaju ili su prekratke)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Neuspesna validacija"
                                        },
                                        "errors": {
                                            "properties": {
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The password confirmation does not match."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kombinacija email-a i tokena nije pronađena u evidenciji",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Neispravan token ili email"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}